refact: dbmodel 수정 및 쿼리 타입설정
