+++ server/src/models/schemas/user.ts
@@ -1,17 +1,18 @@
-import mongoose from "mongoose";
+import mongoose, { Types } from "mongoose";
 import shortId from "./type/shortId.js";
 
-export interface DBUser {
-  shortId: String;
-  email: String;
-  password: String;
-  name: String;
+export interface IUser {
+  _id: Types.ObjectId;
+  // shortId: string;
+  email: string;
+  password: string;
+  name: string;
   status: boolean;
 }
 
-const UserSchema = new mongoose.Schema<DBUser>(
+const UserSchema = new mongoose.Schema<IUser>(
   {
-    shortId,
+    // shortId,
     email: String,
     password: String,
     name: String,
