+++ server/src/routes/index.ts
@@ -1,7 +1,7 @@
 import { Router } from "express";
 import * as users from "./user.routes";
 import * as closet from "./closet.routes";
-// import * as posts from "./posts.routes.js";
+import * as posts from "./post.routes";
 // import * as market from "./market.routes.js";
 // import * as like from "./like.routes.js";
 // import * as movepost from "./movepost.routes.js";
@@ -16,7 +16,7 @@ const router = Router();
 
 router.use(users.path, users.router);
 router.use(closet.path, authmiddleware, closet.router);
-// router.use(posts.path, posts.router);
+router.use(posts.path, posts.router);
 // router.use(market.path, market.router);
 // router.use(like.path, authmiddleware, like.router);
 // router.use(movepost.path, movepost.router);
